requirements.txt
flask
flask-admin
flask-login
flask-sqlalchemy
wtforms-components
pillow
flask-mail
googlemaps
